+++ circus/shared/gui.py
@@ -1040,7 +1040,7 @@ class PreviewGUI(QtGui.QMainWindow):
         self.N_t              = params.getint('data', 'N_t')
         self.spike_thresh     = params.getint('detection', 'spike_thresh')
         self.peaks_sign       = params.get('detection', 'peaks')  
-        self.N_total          = params.getint('data', 'N_total')
+        self.N_total          = numpy.int64(params.getint('data', 'N_total'))
         self.sampling_rate    = params.getint('data', 'sampling_rate')
         self.template_shift   = params.getint('data', 'template_shift')
         self.filename         = params.get('data', 'data_file')
